refactor(client-cli): move mutual functions between download and verify command to shared_steps file
