+++ mithril-client-cli/src/commands/cardano_db/download/mod.rs
@@ -1,5 +1,3 @@
-pub(crate) mod shared_steps;
-
 mod v1;
 mod v2;
