+++ mithril-client-cli/src/commands/cardano_db/verify.rs
@@ -4,27 +4,21 @@ use std::{
     sync::Arc,
 };
 
-use anyhow::{anyhow, Context};
+use anyhow::Context;
 use chrono::Utc;
 use clap::Parser;
-use mithril_client::{
-    cardano_database_client::ImmutableFileRange,
-    common::{ImmutableFileNumber, MKProof, ProtocolMessage},
-    CardanoDatabaseSnapshot, Client, MessageBuilder, MithrilCertificate, MithrilResult,
-};
-use slog::{debug, warn, Logger};
+use mithril_client::MithrilResult;
+use slog::Logger;
 
 use crate::{
-    commands::{cardano_db::CardanoDbCommandsBackend, client_builder, SharedArgs},
-    configuration::{ConfigError, ConfigParameters, ConfigSource},
-    utils::{
-        self, CardanoDbUtils, ExpanderUtils, IndicatifFeedbackReceiver, ProgressOutputType,
-        ProgressPrinter,
+    commands::{
+        cardano_db::{shared_steps, CardanoDbCommandsBackend},
+        client_builder, SharedArgs,
     },
+    configuration::{ConfigError, ConfigParameters, ConfigSource},
+    utils::{self, ExpanderUtils, IndicatifFeedbackReceiver, ProgressOutputType, ProgressPrinter},
     CommandContext,
 };
-
-use super::download::shared_steps;
 
 /// Clap command to verify a Cardano db and its associated certificate.
 #[derive(Parser, Debug, Clone)]
@@ -62,9 +56,6 @@ impl CardanoDbVerifyCommand {
                 r#"The "verify" subcommand is not available for the v1, use --backend v2 instead"#,
             )),
             CardanoDbCommandsBackend::V2 => {
-                // *Important*: All parameters also available in the collected config source below MUST be
-                // obtained using the `params` argument.
-                // No need for `require` as validation of existence will be done by the `NewCardanoDbDownloadCommand` itself.
                 let params = context.config_parameters()?.add_source(self)?;
                 self.verify(context.logger(), params).await
             }
@@ -89,9 +80,7 @@ impl CardanoDbVerifyCommand {
             .with_logger(logger.clone())
             .build()?;
 
-        client
-            .cardano_database_v2()
-            .check_presence_of_immutables(db_dir)?;
+        client.cardano_database_v2().check_has_immutables(db_dir)?;
 
         let get_list_of_artifact_ids = || async {
             let cardano_db_snapshots =
@@ -122,9 +111,9 @@ impl CardanoDbVerifyCommand {
         )
         .await?;
 
-        let immutable_file_range = immutable_file_range(None, None);
+        let immutable_file_range = shared_steps::immutable_file_range(None, None);
 
-        let merkle_proof = compute_verify_merkle_proof(
+        let merkle_proof = shared_steps::compute_verify_merkle_proof(
             2,
             &progress_printer,
             &client,
@@ -135,11 +124,15 @@ impl CardanoDbVerifyCommand {
         )
         .await?;
 
-        let message =
-            compute_cardano_db_snapshot_message(3, &progress_printer, &certificate, &merkle_proof)
-                .await?;
+        let message = shared_steps::compute_cardano_db_snapshot_message(
+            3,
+            &progress_printer,
+            &certificate,
+            &merkle_proof,
+        )
+        .await?;
 
-        verify_cardano_db_snapshot_signature(
+        shared_steps::verify_message_matches_certificate(
             logger,
             4,
             &progress_printer,
@@ -165,15 +158,12 @@ impl CardanoDbVerifyCommand {
         json_output: bool,
     ) -> MithrilResult<()> {
         if json_output {
-            let canonicalized_filepath = &db_dir.canonicalize().with_context(|| {
-                format!(
-                    "Could not get canonicalized filepath of '{}'",
-                    db_dir.display()
-                )
+            let canonical_filepath = &db_dir.canonicalize().with_context(|| {
+                format!("Could not get canonical filepath of '{}'", db_dir.display())
             })?;
             let json = serde_json::json!({
                 "timestamp": Utc::now().to_rfc3339(),
-                "verified_db_directory": canonicalized_filepath
+                "verified_db_directory": canonical_filepath
             });
             println!("{}", json);
         } else {
@@ -204,206 +194,5 @@ impl ConfigSource for CardanoDbVerifyCommand {
         }
 
         Ok(map)
-    }
-}
-
-/// Computes and verifies the Merkle proof for the given certificate and database snapshot.
-pub async fn compute_verify_merkle_proof(
-    step_number: u16,
-    progress_printer: &ProgressPrinter,
-    client: &Client,
-    certificate: &MithrilCertificate,
-    cardano_database_snapshot: &CardanoDatabaseSnapshot,
-    immutable_file_range: &ImmutableFileRange,
-    unpacked_dir: &Path,
-) -> MithrilResult<MKProof> {
-    progress_printer.report_step(step_number, "Computing and verifying the Merkle proof…")?;
-    let merkle_proof = client
-        .cardano_database_v2()
-        .compute_merkle_proof(
-            certificate,
-            cardano_database_snapshot,
-            immutable_file_range,
-            Path::new(&unpacked_dir),
-        )
-        .await?;
-
-    merkle_proof
-        .verify()
-        .with_context(|| "Merkle proof verification failed")?;
-
-    Ok(merkle_proof)
-}
-
-/// Verifies the signature of the Cardano db snapshot against the certificate and message.
-pub async fn verify_cardano_db_snapshot_signature(
-    logger: &Logger,
-    step_number: u16,
-    progress_printer: &ProgressPrinter,
-    certificate: &MithrilCertificate,
-    message: &ProtocolMessage,
-    cardano_db_snapshot: &CardanoDatabaseSnapshot,
-    db_dir: &Path,
-) -> MithrilResult<()> {
-    progress_printer.report_step(step_number, "Verifying the cardano db signature…")?;
-    if !certificate.match_message(message) {
-        debug!(
-            logger,
-            "Merkle root verification failed, removing unpacked files & directory."
-        );
-
-        if let Err(error) = std::fs::remove_dir_all(db_dir) {
-            warn!(
-                logger, "Error while removing unpacked files & directory";
-                "error" => error.to_string()
-            );
-        }
-
-        return Err(anyhow!(
-            "Certificate verification failed (cardano db snapshot hash = '{}').",
-            cardano_db_snapshot.hash.clone()
-        ));
-    }
-
-    Ok(())
-}
-
-/// Constructs an `ImmutableFileRange` based on the provided start and end file numbers.
-pub fn immutable_file_range(
-    start: Option<ImmutableFileNumber>,
-    end: Option<ImmutableFileNumber>,
-) -> ImmutableFileRange {
-    match (start, end) {
-        (None, None) => ImmutableFileRange::Full,
-        (Some(start), None) => ImmutableFileRange::From(start),
-        (Some(start), Some(end)) => ImmutableFileRange::Range(start, end),
-        (None, Some(end)) => ImmutableFileRange::UpTo(end),
-    }
-}
-
-/// Computes the Cardano db snapshot message using the provided certificate and Merkle proof.
-pub async fn compute_cardano_db_snapshot_message(
-    step_number: u16,
-    progress_printer: &ProgressPrinter,
-    certificate: &MithrilCertificate,
-    merkle_proof: &MKProof,
-) -> MithrilResult<ProtocolMessage> {
-    progress_printer.report_step(step_number, "Computing the cardano db snapshot message")?;
-    let message = CardanoDbUtils::wait_spinner(
-        progress_printer,
-        MessageBuilder::new().compute_cardano_database_message(certificate, merkle_proof),
-    )
-    .await
-    .with_context(|| "Can not compute the cardano db snapshot message")?;
-
-    Ok(message)
-}
-
-#[cfg(test)]
-mod tests {
-    use mithril_client::{
-        common::{CardanoDbBeacon, ProtocolMessagePartKey, SignedEntityType},
-        MithrilCertificateMetadata,
-    };
-    use mithril_common::test_utils::TempDir;
-
-    use super::*;
-
-    fn dummy_certificate() -> MithrilCertificate {
-        let mut protocol_message = ProtocolMessage::new();
-        protocol_message.set_message_part(
-            ProtocolMessagePartKey::CardanoDatabaseMerkleRoot,
-            CardanoDatabaseSnapshot::dummy().hash.to_string(),
-        );
-        protocol_message.set_message_part(
-            ProtocolMessagePartKey::NextAggregateVerificationKey,
-            "whatever".to_string(),
-        );
-        let beacon = CardanoDbBeacon::new(10, 100);
-
-        MithrilCertificate {
-            hash: "hash".to_string(),
-            previous_hash: "previous_hash".to_string(),
-            epoch: beacon.epoch,
-            signed_entity_type: SignedEntityType::CardanoDatabase(beacon),
-            metadata: MithrilCertificateMetadata::dummy(),
-            protocol_message: protocol_message.clone(),
-            signed_message: "signed_message".to_string(),
-            aggregate_verification_key: String::new(),
-            multi_signature: String::new(),
-            genesis_signature: String::new(),
-        }
-    }
-
-    #[tokio::test]
-    async fn verify_cardano_db_snapshot_signature_should_remove_db_dir_if_messages_mismatch() {
-        let progress_printer = ProgressPrinter::new(ProgressOutputType::Tty, 1);
-        let certificate = dummy_certificate();
-        let mut message = ProtocolMessage::new();
-        message.set_message_part(
-            ProtocolMessagePartKey::CardanoDatabaseMerkleRoot,
-            "merkle-root-123456".to_string(),
-        );
-        message.set_message_part(
-            ProtocolMessagePartKey::NextAggregateVerificationKey,
-            "avk-123456".to_string(),
-        );
-        let cardano_db = CardanoDatabaseSnapshot::dummy();
-        let db_dir = TempDir::create(
-            "client-cli",
-            "verify_cardano_db_snapshot_signature_should_remove_db_dir_if_messages_mismatch",
-        );
-
-        let result = verify_cardano_db_snapshot_signature(
-            &Logger::root(slog::Discard, slog::o!()),
-            1,
-            &progress_printer,
-            &certificate,
-            &message,
-            &cardano_db,
-            &db_dir,
-        )
-        .await;
-
-        assert!(result.is_err());
-        assert!(
-            !db_dir.exists(),
-            "The db directory should have been removed but it still exists"
-        );
-    }
-
-    #[test]
-    fn immutable_file_range_without_start_without_end_returns_variant_full() {
-        let range = immutable_file_range(None, None);
-
-        assert_eq!(range, ImmutableFileRange::Full);
-    }
-
-    #[test]
-    fn immutable_file_range_with_start_without_end_returns_variant_from() {
-        let start = Some(12);
-
-        let range = immutable_file_range(start, None);
-
-        assert_eq!(range, ImmutableFileRange::From(12));
-    }
-
-    #[test]
-    fn immutable_file_range_with_start_with_end_returns_variant_range() {
-        let start = Some(12);
-        let end = Some(345);
-
-        let range = immutable_file_range(start, end);
-
-        assert_eq!(range, ImmutableFileRange::Range(12, 345));
-    }
-
-    #[test]
-    fn immutable_file_range_without_start_with_end_returns_variant_up_to() {
-        let end = Some(345);
-
-        let range = immutable_file_range(None, end);
-
-        assert_eq!(range, ImmutableFileRange::UpTo(345));
     }
 }
